+++ tests/frontend/org/voltdb/regressionsuites/statistics/TestStatisticsSuiteDatabaseElementStats.java
@@ -150,7 +150,7 @@ public class TestStatisticsSuiteDatabaseElementStats extends StatisticsTestSuite
         expectedSchema[8] = new ColumnInfo("IS_UNIQUE", VoltType.TINYINT);
         expectedSchema[9] = new ColumnInfo("IS_COUNTABLE", VoltType.TINYINT);
         expectedSchema[10] = new ColumnInfo("ENTRY_COUNT", VoltType.BIGINT);
-        expectedSchema[11] = new ColumnInfo("MEMORY_ESTIMATE", VoltType.INTEGER);
+        expectedSchema[11] = new ColumnInfo("MEMORY_ESTIMATE", VoltType.BIGINT);
         VoltTable expectedTable = new VoltTable(expectedSchema);
 
         VoltTable[] results = null;
